refactor(library): Extract component to represent Northern Ireland Assembly elections
