+++ uk_election_timetables/elections/__init__.py
@@ -1,3 +1,4 @@
 from uk_election_timetables.elections.scottish_parliament import *
 from uk_election_timetables.elections.senedd_cymru import *
 from uk_election_timetables.elections.greater_london_assembly import *
+from uk_election_timetables.elections.northern_ireland_assembly import *
